+++ app/views/placements/support/schools/mentors/show.html.erb
@@ -29,5 +29,5 @@
     </div>
   </div>
 
-  <%= govuk_link_to t(".remove_mentor"), remove_placements_support_school_mentor_path(@school, @mentor), class: "app-link app-link--destructive" %>
+  <%= govuk_link_to t(".delete_mentor"), remove_placements_support_school_mentor_path(@school, @mentor), class: "app-link app-link--destructive" %>
 </div>
